Move emailuser model into a separate wagtail.tests.emailuser app
