+++ wagtail/tests/settings.py
@@ -106,6 +106,7 @@ INSTALLED_APPS = (
     'wagtail.tests.testapp',
     'wagtail.tests.demosite',
     'wagtail.tests.customuser',
+    'wagtail.tests.emailuser',
     'wagtail.tests.snippets',
     'wagtail.tests.routablepage',
     'wagtail.tests.search',
